Consolidate CosmosBuilder impls into one file
