+++ crates/cli/cli/src/application/mod.rs
@@ -3,7 +3,7 @@ use std::path::{Path, PathBuf};
 use hermes_cli_framework::application::Application;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 use hermes_runtime::types::runtime::HermesRuntime;
 
 use crate::commands::HermesCommand;
