+++ crates/cli/cli/src/commands/channel/mod.rs
@@ -2,7 +2,7 @@ pub mod create;
 pub use create::ChannelCreate;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 
 use crate::Result;
