+++ crates/cli/cli/src/commands/clear/mod.rs
@@ -1,7 +1,7 @@
 mod packets;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 pub use packets::PacketsClear;
 
 use crate::Result;
