+++ crates/cli/cli/src/commands/client/update.rs
@@ -1,6 +1,6 @@
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 use hermes_cosmos_relayer::contexts::chain::CosmosChain;
 use hermes_error::traits::wrap::ErrorWrapper;
 use hermes_relayer_components::build::traits::components::relay_builder::CanBuildRelay;
