+++ crates/cli/cli/src/commands/connection/mod.rs
@@ -2,7 +2,7 @@ mod create;
 pub use create::ConnectionCreate;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 
 use crate::Result;
