+++ crates/cli/cli/src/commands/keys/list.rs
@@ -3,7 +3,7 @@ use std::collections::BTreeMap;
 
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::{json, Output};
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 use ibc_relayer::config::{ChainConfig, Config};
 use ibc_relayer_types::core::ics24_host::identifier::ChainId;
 use oneline_eyre::eyre::eyre;
