+++ crates/cli/cli/src/commands/keys/mod.rs
@@ -11,7 +11,7 @@ mod balance;
 pub use balance::KeysBalanceCmd;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 
 use crate::Result;
