+++ crates/cli/cli/src/commands/query/channel/mod.rs
@@ -7,7 +7,7 @@ pub use end::QueryChannelEnd;
 pub use ends::QueryChannelEnds;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 
 use crate::Result;
