+++ crates/cli/cli/src/commands/query/client/mod.rs
@@ -8,7 +8,7 @@ mod consensus;
 pub use consensus::QueryClientConsensus;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 
 use crate::Result;
