+++ crates/cli/cli/src/commands/query/clients.rs
@@ -6,7 +6,7 @@ use hermes_cli_components::any_client::types::client_state::AnyClientState;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::{json, Output};
 use hermes_cosmos_chain_components::types::tendermint::TendermintClientState;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 use hermes_cosmos_relayer::contexts::chain::CosmosChain;
 use hermes_error::traits::wrap::ErrorWrapper;
 use hermes_error::types::Error;
