+++ crates/cli/cli/src/commands/query/connection/mod.rs
@@ -2,7 +2,7 @@ mod end;
 pub use end::QueryConnectionEnd;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 
 use crate::Result;
