+++ crates/cli/cli/src/commands/query/mod.rs
@@ -19,7 +19,7 @@ pub use channels::QueryChannels;
 mod packet;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::Output;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 pub use packet::PacketCommands;
 
 use crate::Result;
