+++ crates/cli/cli/src/commands/query/packet/pending.rs
@@ -3,7 +3,7 @@ use core::fmt;
 use hermes_cli_framework::command::CommandRunner;
 use hermes_cli_framework::output::{json, Output};
 use hermes_cosmos_chain_components::traits::chain_handle::HasBlockingChainHandle;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 use ibc_relayer::chain::counterparty::{
     channel_connection_client, channel_on_destination, pending_packet_summary, PendingPackets,
 };
