+++ crates/cosmos/cosmos-relayer/src/contexts/mod.rs
@@ -1,5 +1,5 @@
 pub mod birelay;
-pub mod builder;
+pub mod build;
 pub mod chain;
 pub mod encoding;
 pub mod relay;
