+++ crates/cosmos/cosmos-relayer/src/impls/mod.rs
@@ -1,4 +1,3 @@
-pub mod build;
 pub mod error;
 pub mod subscription;
 pub mod telemetry;
