+++ tools/integration-test/src/tests/context.rs
@@ -1,7 +1,7 @@
 use std::collections::HashMap;
 
 use hermes_cosmos_relayer::contexts::birelay::CosmosBiRelay;
-use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
+use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 use hermes_relayer_components::build::traits::components::birelay_builder::CanBuildBiRelay;
 use hermes_runtime::types::runtime::HermesRuntime;
 use ibc_relayer::chain::handle::ChainHandle;
